feat: update header links and fix window size hook
