+++ src/hooks/useWindowSize.ts
@@ -2,23 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+import { SCREEN_SIZE, pixelWidthToScreenSize } from '@/utils/windowSize';
 import { useEffect, useState } from 'react';
 
-enum SCREEN_SIZE {
-  SM = 'SM',
-  MD = 'MD',
-  LG = 'LG',
-  XL = 'XL',
-}
-
-function pixelWidthToScreenSize(width: number): SCREEN_SIZE {
-  if (width < 768) return SCREEN_SIZE.SM;
-  else if (width < 1024 && width >= 768) return SCREEN_SIZE.MD;
-  else if (width < 1280 && width >= 1024) return SCREEN_SIZE.LG;
-  return SCREEN_SIZE.XL;
-}
-
-function useWindowSize() {
+function useWindowSize(): SCREEN_SIZE {
   const [windowSize, setWindowSize] = useState({
     width: 1024,
     height: 1024,
